+++ biweeklybudget/tests/acceptance/test_db_event_handlers.py
@@ -195,11 +195,11 @@ class TestIsFieldsSet(AcceptanceHelper):
         txn = OFXTransaction(
             account=acct,
             statement=stmt,
-            fitid='BankOne.9.1',
+            fitid='BankOne-9-1',
             trans_type='Credit',
             date_posted=stmt.ledger_bal_as_of,
             amount=Decimal('1234.56'),
-            name='BankOne.9.1'
+            name='BankOne-9-1'
         )
         testdb.add(txn)
         testdb.commit()
@@ -211,11 +211,11 @@ class TestIsFieldsSet(AcceptanceHelper):
         txn = OFXTransaction(
             account=acct,
             statement=stmt,
-            fitid='BankOne.9.2',
+            fitid='BankOne-9-2',
             trans_type='Credit',
             date_posted=stmt.ledger_bal_as_of,
             amount=Decimal('1234.56'),
-            name='re-other-fee BankOne.9.2'
+            name='re-other-fee BankOne-9-2'
         )
         testdb.add(txn)
         testdb.commit()
@@ -227,11 +227,11 @@ class TestIsFieldsSet(AcceptanceHelper):
         txn = OFXTransaction(
             account=acct,
             statement=stmt,
-            fitid='BankOne.9.3',
+            fitid='BankOne-9-3',
             trans_type='Credit',
             date_posted=stmt.ledger_bal_as_of,
             amount=Decimal('1234.56'),
-            name='Late Fee BankOne.9.3'
+            name='Late Fee BankOne-9-3'
         )
         testdb.add(txn)
         testdb.commit()
@@ -243,11 +243,11 @@ class TestIsFieldsSet(AcceptanceHelper):
         txn = OFXTransaction(
             account=acct,
             statement=stmt,
-            fitid='BankOne.9.4',
+            fitid='BankOne-9-4',
             trans_type='Credit',
             date_posted=stmt.ledger_bal_as_of,
             amount=Decimal('1234.56'),
-            name='payment BankOne.9.4'
+            name='payment BankOne-9-4'
         )
         testdb.add(txn)
         testdb.commit()
@@ -259,11 +259,11 @@ class TestIsFieldsSet(AcceptanceHelper):
         txn = OFXTransaction(
             account=acct,
             statement=stmt,
-            fitid='BankOne.9.5',
+            fitid='BankOne-9-5',
             trans_type='Credit',
             date_posted=stmt.ledger_bal_as_of,
             amount=Decimal('1234.56'),
-            name='Thank You BankOne.9.5'
+            name='Thank You BankOne-9-5'
         )
         testdb.add(txn)
         testdb.commit()
@@ -275,7 +275,7 @@ class TestIsFieldsSet(AcceptanceHelper):
         txn = OFXTransaction(
             account=acct,
             statement=stmt,
-            fitid='BankOne.9.6',
+            fitid='BankOne-9-6',
             trans_type='Credit',
             date_posted=stmt.ledger_bal_as_of,
             amount=Decimal('1234.56'),
@@ -288,3 +288,68 @@ class TestIsFieldsSet(AcceptanceHelper):
         assert txn.is_interest_charge is False
         assert txn.is_other_fee is False
         assert txn.is_interest_payment is False
+
+    def test_1_account_re_change_triggers_update_is_fields(self, testdb):
+        """
+        Test for
+        :py:func:`~.db_event_handlers.handle_account_re_change`
+        """
+        acct = testdb.query(Account).get(1)
+        acct.re_interest_paid = None
+        testdb.commit()
+        assert acct.re_interest_charge == '^interest-charge'
+        assert acct.re_interest_paid is None
+        assert acct.re_payment == '^(payment|thank you)'
+        assert acct.re_late_fee == '^Late Fee'
+        assert acct.re_other_fee == '^re-other-fee'
+        stmt = testdb.query(OFXStatement).get(1)
+        assert stmt.account_id == 1
+        assert stmt.filename == '/stmt/BankOne/0'
+        txn1 = testdb.query(OFXTransaction).get((1, 'BankOne-9-1'))
+        assert txn1.name == 'BankOne-9-1'
+        assert txn1.is_payment is False
+        assert txn1.is_late_fee is False
+        assert txn1.is_interest_charge is False
+        assert txn1.is_other_fee is False
+        assert txn1.is_interest_payment is False
+        txn2 = testdb.query(OFXTransaction).get((1, 'BankOne-9-2'))
+        assert txn2.name == 're-other-fee BankOne-9-2'
+        assert txn2.is_payment is False
+        assert txn2.is_late_fee is False
+        assert txn2.is_interest_charge is False
+        assert txn2.is_other_fee is True
+        assert txn2.is_interest_payment is False
+        txn3 = testdb.query(OFXTransaction).get((1, 'BankOne-9-3'))
+        assert txn3.name == 'Late Fee BankOne-9-3'
+        assert txn3.is_payment is False
+        assert txn3.is_late_fee is True
+        assert txn3.is_interest_charge is False
+        assert txn3.is_other_fee is False
+        assert txn3.is_interest_payment is False
+        # change the account
+        acct.re_interest_charge = '^BankOne-9-1$'
+        acct.re_payment = '^Late Fee'
+        acct.re_late_fee = '^foobarbaz'
+        testdb.commit()
+        # re-confirm
+        txn1 = testdb.query(OFXTransaction).get((1, 'BankOne-9-1'))
+        assert txn1.name == 'BankOne-9-1'
+        assert txn1.is_payment is False
+        assert txn1.is_late_fee is False
+        assert txn1.is_interest_charge is True
+        assert txn1.is_other_fee is False
+        assert txn1.is_interest_payment is False
+        txn2 = testdb.query(OFXTransaction).get((1, 'BankOne-9-2'))
+        assert txn2.name == 're-other-fee BankOne-9-2'
+        assert txn2.is_payment is False
+        assert txn2.is_late_fee is False
+        assert txn2.is_interest_charge is False
+        assert txn2.is_other_fee is True
+        assert txn2.is_interest_payment is False
+        txn3 = testdb.query(OFXTransaction).get((1, 'BankOne-9-3'))
+        assert txn3.name == 'Late Fee BankOne-9-3'
+        assert txn3.is_payment is True
+        assert txn3.is_late_fee is False
+        assert txn3.is_interest_charge is False
+        assert txn3.is_other_fee is False
+        assert txn3.is_interest_payment is False